% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-es.R
\name{convert_r_d}
\alias{convert_r_d}
\alias{convert_tau_r}
\title{Convert various effect size measures}
\usage{
convert_r_d(r, nA, nB)

convert_tau_r(tau)
}
\arguments{
\item{r}{Pearson correlation}

\item{nA}{Measurement times phase A}

\item{nB}{Measurement times phase B}

\item{tau}{Kendall's tau}
}
\description{
Convert various effect size measures
}
