% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_functions.R
\name{get_mcfn}
\alias{get_mcfn}
\title{Get a predefined function for the Monte Carlo procedure}
\usage{
get_mcfn(x)
}
\arguments{
\item{x}{Character with the name of a predefined function. Leave empty to
get the names of all possible functions.}
}
\value{
A function or a list of function names.
}
\description{
Get a predefined function for the Monte Carlo procedure
}
