% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_functions.R
\docType{data}
\name{mcfn}
\alias{mcfn}
\title{List Object with predefined functions.}
\format{
An object of class \code{list} of length 15.
}
\usage{
mcfn
}
\description{
List Object with predefined functions.
}
\keyword{datasets}
