% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcplot.R
\name{mcplot}
\alias{mcplot}
\title{MC Plot}
\usage{
mcplot(
  data_mc,
  line_curved = FALSE,
  add_points = TRUE,
  caption = FALSE,
  var_x = 1,
  var_shape = 2,
  var_facet = "Method",
  var_col = 3,
  ncol = 2,
  reverse = FALSE,
  marks = c(5, 80),
  ylim = c(0, 100),
  ylab = "Percentage",
  labels_col = NULL
)
}
\description{
MC Plot
}
