% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcscan.R
\name{mcscan}
\alias{mcscan}
\title{MC scan}
\usage{
mcscan(
  design,
  iterations,
  method,
  parameter = "p",
  alpha_test = TRUE,
  power_test = TRUE,
  store_design = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
MC scan
}
