% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcstudy.R
\name{mcstudy}
\alias{mcstudy}
\title{Design frame for a Monte-Carlo study}
\usage{
mcstudy(
  design,
  iterations,
  methods,
  store_design = FALSE,
  rf_arguments = NULL,
  n_sims = 100,
  ...
)
}
\arguments{
\item{design}{A design template created with the design_template() function}

\item{iterations}{A list with variables and their iterations.}

\item{methods}{A list of functions that analyze the random datasets.}

\item{store_design}{If true, the random datasets are stroed in the returned object.}

\item{...}{Further paramters passed to the mcscan() function.}
}
\description{
Design frame for a Monte-Carlo study
}
