% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctable.R
\name{mctable}
\alias{mctable}
\title{Result table for mcstudy}
\usage{
mctable(
  data_mc,
  wider = NULL,
  format = "df",
  digits = 2,
  var_y = "p",
  label_header = NULL
)
}
\arguments{
\item{data_mc}{Return from the mcstudy function}
}
\description{
Result table for mcstudy
}
