% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcplot.R
\name{mcplot}
\alias{mcplot}
\title{Monte Carlo plot}
\usage{
mcplot(
  data_mc,
  add_points = TRUE,
  caption = FALSE,
  var_y = "p",
  var_x = 1,
  var_shape = 2,
  var_facet = "Methods",
  var_col = 3,
  ncol = 2,
  marks = NULL,
  ylim = NULL,
  ylab = "Value",
  labels_col = NULL,
  template = NULL
)
}
\arguments{
\item{data_mc}{An object returned from the mcstudy() function.}

\item{add_points}{If TRUE, lines get a plot.}

\item{caption}{TRUE or FALSE.}

\item{var_x}{Name of the variable to be placed on the x axis.}

\item{var_shape}{Name of the variable to be depicted in point shapes.}

\item{var_facet}{Name of the variable to create facets.}

\item{ncol}{Number of columns for the facets.}

\item{marks}{Vector with values for adding horizontal lines.}

\item{ylim}{Vector with two values defining the limits for the y axis.}

\item{ylab}{Character string with the y axis label.}

\item{labels_col}{Labels for the color legend.}
}
\description{
Monte Carlo plot
}
